+++ eds-aic-20230930/src/Models/CreatePolicyGroupRequest/netRedirectPolicy/rules.php
@@ -1,0 +1,59 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Edsaic\V20230930\Models\CreatePolicyGroupRequest\netRedirectPolicy;
+
+use AlibabaCloud\Tea\Model;
+
+class rules extends Model
+{
+    /**
+     * @var string
+     */
+    public $ruleType;
+
+    /**
+     * @var string
+     */
+    public $target;
+    protected $_name = [
+        'ruleType' => 'RuleType',
+        'target'   => 'Target',
+    ];
+
+    public function validate()
+    {
+    }
+
+    public function toMap()
+    {
+        $res = [];
+        if (null !== $this->ruleType) {
+            $res['RuleType'] = $this->ruleType;
+        }
+        if (null !== $this->target) {
+            $res['Target'] = $this->target;
+        }
+
+        return $res;
+    }
+
+    /**
+     * @param array $map
+     *
+     * @return rules
+     */
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['RuleType'])) {
+            $model->ruleType = $map['RuleType'];
+        }
+        if (isset($map['Target'])) {
+            $model->target = $map['Target'];
+        }
+
+        return $model;
+    }
+}